requirements.txt
#  python=3.10
# 我尝试了40系列的显卡，这说明30系列的显卡可能也适用
# torch==1.10.1，CUDA=11.3
# torch==2.0.1，CUDA=11.8

# 在50系列显卡（可能更高级的显卡）
# torch=2.7.1，CUDA=12.8

tqdm==4.66.1
Pillow==10.0.1
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.5
pycocotools>=2.0.6

# -i https://pypi.douban.com/simple/
# -i https://pypi.tuna.tsinghua.edu.cn/simple
# -i https://mirrors.aliyun.com/pypi/simple/